refactor: make SimpleUploadController simpler
